Fix the click and release functions. Refactor draw pixels methods. Mouse cursor use util's bit handling functions. Fix bit handling functions - change to 32 bit variable instead of just 8 bit.

+++ util/bit_handling.c
@@ -1,16 +1,16 @@
 #include <stdint.h>
 #include "../cpu/types.h"
 
-boolean is_bit_set(uint8_t byte_value, uint8_t bit_index_0_based){
+boolean is_bit_set(uint32_t byte_value, uint8_t bit_index_0_based){
 	return byte_value & 1 << bit_index_0_based;
 }
 
-uint8_t set_bit(uint8_t byte_value, uint8_t bit_index_0_based){
+uint8_t set_bit(uint32_t byte_value, uint8_t bit_index_0_based){
 	byte_value |= 1 << bit_index_0_based;
 	return byte_value;
 }
 
-uint8_t clear_bit (uint8_t byte_value, uint8_t bit_index_0_based){
+uint8_t clear_bit (uint32_t byte_value, uint8_t bit_index_0_based){
 	byte_value &= ~(1<<bit_index_0_based);
 	return byte_value;
 }
